+++ data/data.h
@@ -7,7 +7,7 @@ struct person_t {
     std::string name;
     std::string lastName;
     std::string email;
-    uint8_t age;
+    int16_t age;
     int id;
 };
 
@@ -22,7 +22,7 @@ class user {
         void setName( std::string name);
         void setLastName( std::string name);
         void setEmail( std::string name);
-        void setAge( uint8_t setAge);
+        void setAge( int16_t setAge);
         void setId( int id);
         void setVariables( const struct person_t obj);
